Mummi and bilbo contribution!! veriverse contribution!! LCTV contribution!! catherinacortega, sjpablo (Pawel Kulesza), fuel (Benjamin Kuhl),  mushmouth (Patrick McFarling), sonnystar (sonnystaminecraft, Vlad Ionescu), abbyyyyy, alexcode, ayvatan, blackhatisme, boza, chrifrachid, coderforever,cookiesliyr, cubor, denvercowboy, diagram, dzetko, euphemisminc, gillie, gzj71ck4elm2, hudwan, intelijens (Michael Bui), LaffayetteBrittoPhoto (streamer with its C300 photo), jarzi, johnyblaze, jookeer, kaeris, kissloryshy, kord, korionintell..., letscodenow, loci, lyndroid, magrib, marianyc, michaeljcalkins, nuke, paezrice, peter_cados, phanxgames (Henry Price), purpleizlegitx, rheghal, rooney2, silvaesouza, siox, sova, systemis, thiagovgamg, usandy, v0n3x, vujadinovicdejan, wanket, wapaca, zaaz and zakaria98 contribution!! jiangyubo530, kozmo, contribution!! Twitch contribution!! AlfredoPlays (and his mates), Bialyorzeu, cpmgamingltd, GameDevCompany (Björn and Arash contribution...), GrubbyJuicde, Handmadehero (Casey Muratori), Mahakhef, Mattyplays6, michaeljcalkins, nothings, Preid2404, Stavros__, Tectuma (GothicatCat1(Stacie), Tectuma(Kevin), Matty and Dylan) and Xandergamingtwich  contribution!! Beam.pro contribution!! BenB2310 contribution!! bisba contribution!! caro contribution!!
bilbo at veriverse
i7 6700 32GB RAM Ubuntu 16.04 Unity Numix Theme Numix Circle Icons
(booting also mac OS Sierra and Windows 10 Pro/Hyper-V)

+++ appseed/aura/primitive/primitive_var.cpp
@@ -3618,7 +3618,7 @@ var::operator bool() const
     || m_etype == type_filetime
     || m_etype == type_time)
    {
-      
+
       return true;
 
    }
@@ -3660,7 +3660,7 @@ var::operator bool() const
    }
    else if(m_etype == type_string)
    {
-      
+
       return m_str.CompareNoCase("true") == 0;
 
    }
@@ -3803,7 +3803,7 @@ bool var::is_property_true(const char * pszName) const
 
 bool var::is_property_false(const char * pszName) const
 {
-   
+
    return !is_property_true(pszName);
 
 }
@@ -4008,7 +4008,7 @@ end:
    }
    else
    {
-      operator = (atoll(string(pszStart, pszParse - pszStart)));
+      operator = ((int64_t)atoll(string(pszStart, pszParse - pszStart)));
    }
    psz = pszParse;
 }
@@ -4184,7 +4184,7 @@ void var::parse_json(const char * & pszJson, const char * pszEnd)
    {
 
       ::str::consume_spaces(pszJson, 0, pszEnd);
-      
+
       if (*pszJson == '\0')
       {
 
@@ -4214,7 +4214,7 @@ void var::parse_json(const char * & pszJson, const char * pszEnd)
 
          if (varChild.get_string().CompareNoCase(id) == 0)
          {
-            
+
             ::str::consume_spaces(pszJson, 0, pszEnd);
             ::str::consume(pszJson, ":", 1, pszEnd);
 
@@ -4239,11 +4239,11 @@ void var::parse_json(const char * & pszJson, const char * pszEnd)
          }
          else
          {
-            
+
             string str = "not expected character : ";
-         
+
             str += pszJson;
-            
+
             throw_json_parsing_exception(str);
 
          }
@@ -4292,7 +4292,7 @@ void var::parse_json(const char * & pszJson, const char * pszEnd)
       if (*pszJson == ']')
       {
          pszJson++;
-         
+
          return etype;
 
       }
@@ -4363,7 +4363,7 @@ void var::parse_json(const char * & pszJson, const char * pszEnd)
    }
    else if (*pszJson == '[')
    {
-      
+
       return var::type_new;
 
    }
@@ -4590,7 +4590,7 @@ bool var::is_false() const
 
    if (get_type() != type_bool)
    {
-      
+
       return false;
 
    }
